backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
yfinance==0.2.43
prophet==1.1.5
statsmodels==0.14.2
pandas==2.2.2
numpy==2.1.1
scikit-learn==1.5.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-multipart==0.0.9
supabase==2.7.2
httpx==0.27.2
pydantic==2.9.1
python-dotenv==1.0.1
